requirements.txt
Flask==0.10.1
Jinja2==2.8
MarkupSafe==0.23
Twisted==16.1.1
Werkzeug==0.11.8
argparse==1.2.1
itsdangerous==0.24
six==1.10.0
wsgiref==0.1.2
zope.interface==4.1.3